requirements.txt
blinker==1.8.2
boto3==1.34.129
botocore==1.34.129
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.2.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-PyMongo==2.3.0
fsspec==2024.6.0
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
PyJWT==2.8.0
pymongo==4.7.3
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.1
reportlab==4.2.2
requests==2.32.3
s3fs==0.4.2
s3transfer==0.10.1
six==1.16.0
stripe==10.3.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3